Replace unrecognized characters (that Arduino IDE complains about).
It looks like prior to 95a0f62 the Greek 'mu' character was used to denote 'micro', as in microamps and microwatts. The specific character was Unicode 'MICRO SIGN' (U+00B5). Something in the process of creating that commit changed instances of 'mu' into unrecognized characters.

Some instances were later turned in to Latin/English lowercase 'u', for example in ee1a73b.

This change updates remaining instances of the unrecognized character to 'u'.

Fixes #36